+++ src/components/registration/registration.vue
@@ -163,6 +163,23 @@
             </div>
 
             <div
+              v-if="accountApprovalRequired"
+              class="form-group"
+            >
+              <label
+                class="form--label"
+                for="reason"
+              >{{ $t('registration.reason') }}</label>
+              <textarea
+                id="reason"
+                v-model="user.reason"
+                :disabled="isPending"
+                class="form-control"
+                :placeholder="reasonPlaceholder"
+              />
+            </div>
+
+            <div
               v-if="captcha.type != 'none'"
               id="captcha-group"
               class="form-group"